% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference() is the function to conduct simultaneous inference of estimated partial correlations.}
\usage{
inference(list, alpha = 0.05)
}
\arguments{
\item{list}{A list returned by either `PCGII()` or `clevel()`.}

\item{alpha}{A pre-determined False Discovery Rate. Nominal FDR is set at 0.05 by default.}
}
\value{
An adjacency matrix of significant partial correlations.
}
\description{
Inference() is the function to conduct simultaneous inference of estimated partial correlations.
}
\examples{
PCGII_out=PCGII(df=X, prior=as.data.frame(prior_set), lambda = lam)
sigs_mat=inference(PCGII_out, alpha=.1)
}
